refactor: egoist.components.runtimecontext -> egoist.runtime
